requirements.txt
Flask==2.3.2
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0
openai
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.3
Flask-Caching==2.1.0
redis==5.0.1